fastapi+nextjs/backend/requirements.txt
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.6.0
fastapi==0.115.0
idna==3.10
iso8601==1.1.0
pydantic==2.9.2
pydantic_core==2.23.4
pypika-tortoise==0.1.6
pytz==2024.2
setuptools==75.1.0
sniffio==1.3.1
starlette==0.38.6
tortoise-orm==0.21.6
typing_extensions==4.12.2
wheel==0.44.0
